tests/wpt/tests/html/semantics/interactive-elements/the-dialog-element/dialog-synthetic-keydown.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<body>
<dialog id="dialog"></dialog>
<script>
test(() => {
  dialog.showModal();

  let keydown = new KeyboardEvent("keydown", {key: "Escape", keyCode: 27});
  document.body.dispatchEvent(keydown);
  let keyup = new KeyboardEvent("keyup", {key: "Escape", keyCode: 27});
  document.body.dispatchEvent(keyup);

  assert_true(dialog.open);

  let keyup2 = document.createEvent("Event");
  keyup2.initEvent("keyup", true);
  document.body.dispatchEvent(keyup2);

  assert_true(dialog.open);
}, "close via synthesized Esc key must not work");
</script>
